remove go-msgpack from benchmarks
go-msgpack has been stalled and has been deprecated for over a year.
The code was moved into go-codec and all development continued there.
It should not be reflected in this benchmark.

Also, use AsSymbolNone for Binc.
Binc is either optimized for space (de-duplicate strings) or speed (do not de-duplicate strings).
For performance comparisons, do not de-duplicate strings; this is in line with most other impls.