report: Initial version
This adds the initial version of a dynamic benchmark results report.

The report shows both marhsal and unmarshal metrics individually and
aggregated. It also records how the serialization of time.Time values is
done and what is the general coding paradigm used by code that intends
to use the serializer.

The report is generated as an HTML report and uses ParamQuery's Grid
component to provide a better UX.