multi: Note benchmarks that perform buffer reuse
This adds an explicit note for the benchmarks which perform buffer reuse
during marshalling operations. This improves fairness when comparing the
performance of different benchmarks that may or may not offer this
feature.

Given that buffer reuse is a marshalling feature and unsafe string
decoding is an unmarshalling feature, for serializers that support both
features, they are both coalesced into a single benchmark to avoid
unnecessarily increasing the size of the report.