main: Verify benchmark finished when validating
Unfortunately, benchmarks ran with testing.Benchmark() do not return an
error, panic or otherwise make their calling goroutine exit early in
order to signal failure.

This could cause validation errors to be silently ignored when running
with -validate specified.

This commit addresses this issue by using a flag to detect when the test
was prematurely finished to report it as an error, with a message for
the user to re-run the test in a way that can actually report the error.