Add baseline_rw benchmark (#177)
* baseline: Add Reader/Writer variant

This variant is meant to demonstrate and generate a baseline for
serializers that can marshal/unmarshal to/from standard io Reader and
Writer intefaces.

While the decision on reusing auxiliary buffers is arbitrary, it is
assumed that writing native types (uint64, etc) can be easily done by
any Reader/Writer implementation itself (or their callers).

The ultimate goal for this baseline test is as a benchmark against other
implementations that offer this method of serialization, strictly
measuring only the overhead of going through the io interfaces.

Both safe and unsafe+reuse implementations are provided.

* report: Update with baseline_rw results