Fix -validate and add -name flags (#176)
* main: Add -name flag

This flag allows selecting a subset of benchmarks to run when generating
the report or validating benchmarks.

This is useful during development in order to verify only newly added
benchmarks as opposed to all benchmarks.

* main: Verify benchmark finished when validating

Unfortunately, benchmarks ran with testing.Benchmark() do not return an
error, panic or otherwise make their calling goroutine exit early in
order to signal failure.

This could cause validation errors to be silently ignored when running
with -validate specified.

This commit addresses this issue by using a flag to detect when the test
was prematurely finished to report it as an error, with a message for
the user to re-run the test in a way that can actually report the error.

* goserbench: Add validation for Money field

This validation was missing from the code.